requirements.txt
pyTelegramBotAPI
peewee
requests
python-dotenv